+++ sites/svelte.dev/src/lib/utils/compat.js
