+++ sites/svelte.dev/src/lib/utils/events.js
