+++ sites/svelte.dev/src/lib/utils/highlight.js
